Update the Makefile and docs since make update-deps doesn't work
The make update-deps command was using glide, but we have moved
to using go mod, so remove the make update-deps command.
Add steps to the README as well on how to update deps and code
generation.
make generate fails if the go install command in the update-codegen.sh
script doesn't have "--mod=vendor" set.

Signed-off-by: Urvashi Mohnani <umohnani@redhat.com>