add mco applyconfiguration hack, mco to codegen
The MCO's API unfortunately has embedded a corev1.ObjectReference for
years. As part of the MCO's API migration, we tried to generate clients,
but this resulted in broken applyconfigurations for the MCO, because
applyconfiguration-gen doesn't regard the corev1.ObjectReference
as an external type without being explicitly specified, so when it
calculates "reachability" of the type, it's not in the list.

This commit adds machineconfiguration to the list of packages for
generation, and also temporarily splits the MCO's
applyconfiguration-gen commands out into a special MCO-only block so it
can reference corev1 as external and generate properly without
affecting generation for the other packages here -- at least until the
MCO can get rid of that embedded reference.