Don't hardcode vendor directories in hack/update-codegen.sh
Get the current location of the openshift/api and code-generator
packages from go list instead of assuming they are in the vendor
directory.

When vendoring is in use, which it normally is, this will continue to
return the correct vendor directory. However, when it has been
overridden, for example by using a temporary go.work during local
development, it will automatically use the overridden location.