Add openshift/api explicitly to dependency magnet
This fixes a bootstrapping issue in update-codegen.sh. Currently the
only dependency in openshift/client-go on a specific api, e.g.
openshift/api/helm/v1beta1, is from generated code. Because packages are
only vendored if they are used, this means that until we generate code
for a package, that package will not be vendored. But it will not be
vendored unless we have a dependency on it from generating code.

You can reproduce this either by trying to add a new openshift/api
package to openshift/client-go, or by:
* find helm -type f -delete
* go mod vendor
* ./hack/update-codegen.sh

The go mod vendor will cause the helm package to be deleted from the
vendor directory because it is no longer referenced, so
update-codegen.sh will not restore it.

We fix this by adding the openshift/api package to the dependency magnet
explicitly. This works because this package contains install.go which in
turn pulls in every api package.